ajustes

+++ src/Alimentacao.java
@@ -2,7 +2,7 @@ public class Alimentacao extends Loja {
 
     private Data dataAlvara;
 
-    public Alimentacao(String nome, int quantidadeFuncionarios, int salarioBaseFuncionario,
+    public Alimentacao(String nome, int quantidadeFuncionarios, double salarioBaseFuncionario,
                        Endereco endereco, Data dataFundacao, Data dataAlvara, int quantidadeEstoque) {
         super(nome, quantidadeFuncionarios, salarioBaseFuncionario, endereco, dataFundacao, quantidadeEstoque);
         this.dataAlvara = dataAlvara;
