+++ src/Cosmetico.java
@@ -2,7 +2,7 @@ public class Cosmetico extends Loja {
 
     private double taxaComercializacao;
 
-    public Cosmetico(String nome, int quantidadeFuncionarios, int salarioBaseFuncionario,
+    public Cosmetico(String nome, int quantidadeFuncionarios, double salarioBaseFuncionario,
                      Endereco endereco, Data dataFundacao, double taxaComercializacao, int quantidadeEstoque) {
         super(nome, quantidadeFuncionarios, salarioBaseFuncionario, endereco, dataFundacao, quantidadeEstoque);
         this.taxaComercializacao = taxaComercializacao;
